feat: added parametrized contract and adapted the code
