+++ README.md
@@ -20,7 +20,7 @@ The use cases implemented in this repository are based on the research paper [Sm
 The 21 use cases identified in the research paper are as follows:
 
 1. [Bet](bet/README.md)  
-2. Simple transfer  
+2. [Simple transfer](simple-transfer.README.md)  
 3. Token transfer  
 4. HTLC  
 5. [Escrow](escrow/README.md)
