--- simple-transfer/offchain/ccl-java/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Thu Jul 31 16:54:12 CEST 2025
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-8.8-bin.zip
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
