--- simple-transfer/offchain/ccl-java/settings.gradle.kts
@@ -1,2 +1,0 @@
-rootProject.name = "ccl-java"
-
